glossarist/concept-model#15 Moved uml classes to the root folder, removed additional namespace
