+++ lib/lutaml/uml.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'lutaml/uml/version'
-require "lutaml/uml/interface/command_line"
+require 'lutaml/uml/interface/command_line'
 
 module Lutaml
   module Uml
