+++ lib/lutaml/uml/constraint.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+##
+## Behaviour metamodel
+##
+module Lutaml
+  module Uml
+    class Constraint < TopElement
+      attr_accessor :body
+    end
+  end
+end
