+++ lib/lutaml/uml/final_state.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+##
+## Behaviour metamodel
+##
+module Lutaml
+  module Uml
+    class FinalState < State
+    end
+  end
+end
