+++ lib/lutaml/uml/interface/base.rb
@@ -13,9 +13,11 @@ module Lutaml
 
         include HasAttributes
 
+        # rubocop:disable Rails/ActiveRecordAliases
         def initialize(attributes = {})
-          update(attributes)
+          update_attributes(attributes)
         end
+        # rubocop:enable Rails/ActiveRecordAliases
 
         def run
           raise NotImplementedError
