+++ lib/lutaml/uml/model.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Lutaml
+  module Uml
+    class Model < Package
+      attr_accessor :viewpoint
+
+      def initialize
+        @contents = []
+      end
+    end
+  end
+end
