+++ lib/lutaml/uml/port.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Lutaml
+  module Uml
+    class Port < Property
+    end
+  end
+end
