--- lib/lutaml/uml/representers/classifier.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Lutaml
-  module Uml
-    module Representers
-      class Classifier < TopElement
-        attr_accessor :generalization
-      end
-    end
-  end
-end
