--- lib/lutaml/uml/representers/connector.rb
@@ -1,23 +1,0 @@
-# frozen_string_literal: true
-
-##
-## Behaviour metamodel
-##
-module Lutaml
-  module Uml
-    module Representers
-      class Connector < TopElement
-        attr_accessor :kind, :connector_end
-
-        def initialize
-          @name = nil
-          @xmi_id = nil
-          @xmi_uuid = nil
-          @connector_end = []
-          @namespace = nil
-          @kind = nil
-        end
-      end
-    end
-  end
-end
