--- lib/lutaml/uml/representers/instance.rb
@@ -1,19 +1,0 @@
-# frozen_string_literal: true
-
-module Lutaml
-  module Uml
-    module Representers
-      class Instance < TopElement
-        attr_accessor :classifier, :slot
-
-        def initialize
-          @name = nil
-          @xmi_id = nil
-          @xmi_uuid = nil
-          @classifier = nil
-          @slot = []
-        end
-      end
-    end
-  end
-end
